requirements.txt
torch
torchvision
opencv-python
tensorboard
tensorflow
numpy
scipy
sqlitedict
tqdm
psutil
tensorboardX
pandas
matplotlib
mongo
asyncio-nats-client
pysnooper
compress_pickle
flask
scikit-learn
terminaltables
peewee
psycopg2
flask-restful
requests
xlrd
statsmodels
redis
pyyml
fire
pillow==6.1